Тесты

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -7,13 +7,11 @@ import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
-@Slf4j
 @Service
+@Slf4j
 public class UserService {
     private final UserStorage userStorage;
 
@@ -21,75 +19,57 @@ public class UserService {
         this.userStorage = userStorage;
     }
 
-    public void addFriend(int acceptorId, int requesterId) throws NotFoundException {
-        User acceptor = getUserOrThrow(acceptorId);
-        User requester = getUserOrThrow(requesterId);
-
-        if (!acceptor.getIncomingFriendRequests().contains(requesterId)) {
-            throw new NotFoundException("Запрос на дружбу не найден");
-        }
-
-        userStorage.acceptFriendRequest(acceptorId, requesterId);
-
-        User updatedAcceptor = userStorage.findById(acceptorId);
-        User updatedRequester = userStorage.findById(requesterId);
-
-        acceptor.setIncomingFriendRequests(updatedAcceptor.getIncomingFriendRequests());
-        acceptor.setFriends(updatedAcceptor.getFriends());
-        requester.setOutgoingFriendRequests(updatedRequester.getOutgoingFriendRequests());
-        requester.setFriends(updatedRequester.getFriends());
-
-        acceptor.getIncomingFriendRequests().remove(requesterId);
-        requester.getOutgoingFriendRequests().remove(acceptorId);
-    }
-
-
-    public void friendShipOffer(int senderId, int recipientId) throws ValidationException, NotFoundException {
-        User sender = getUserOrThrow(senderId);
-        User recipient = getUserOrThrow(recipientId);
-
-        if (sender.getFriends().contains(recipientId)) {
-            throw new ValidationException("Пользователи уже друзья");
-        }
-
-        if (sender.getOutgoingFriendRequests().contains(recipientId)) {
-            throw new ValidationException("Заявка уже отправлена");
-        }
-
-        userStorage.sendFriendRequest(senderId, recipientId);
-
-        User updatedSender = userStorage.findById(senderId);
-        User updatedRecipient = userStorage.findById(recipientId);
-
-        sender.setOutgoingFriendRequests(updatedSender.getOutgoingFriendRequests());
-        recipient.setIncomingFriendRequests(updatedRecipient.getIncomingFriendRequests());
-    }
-
-
-    public void removeFriend(int userId, int friendId) throws NotFoundException {
+    public void addFriend(int userId, int friendId) throws NotFoundException, ValidationException {
         User user = getUserOrThrow(userId);
         User friend = getUserOrThrow(friendId);
 
-        if (!user.getFriends().contains(friendId)) {
-            throw new NotFoundException("Друг не найден");
+        if (user.getFriends().contains(friendId)) {
+            throw new ValidationException("Пользователи уже друзья");
         }
 
-        userStorage.removeFriend(userId, friendId);
-        user.getFriends().remove(friendId);
-        friend.getFriends().remove(userId);
+        // Добавляем только одностороннюю дружбу (user -> friend)
+        userStorage.addFriend(userId, friendId);
+        user.getFriends().add(friendId);
+
+        log.info("Пользователь {} добавил пользователя {} в друзья", userId, friendId);
     }
 
     public List<User> getFriends(int userId) throws NotFoundException {
         User user = getUserOrThrow(userId);
-        return user.getFriends().stream().map(userStorage::findById).collect(Collectors.toList());
+        System.out.println("userId=" + userId + ", friends=" + user.getFriends());
+        return user.getFriends().stream()
+                .map(id -> {
+                    try {
+                        return getUserOrThrow(id);
+                    } catch (NotFoundException e) {
+                        return null;
+                    }
+                })
+                .filter(Objects::nonNull)
+                .collect(Collectors.toList());
+    }
+
+
+    public boolean removeFriend(int userId, int friendId) throws NotFoundException {
+        User user = getUserOrThrow(userId);
+        User friend = getUserOrThrow(friendId);
+        friend.getFriends().remove(userId);
+        user.getFriends().remove(friendId);
+        userStorage.removeFriend(userId, friendId);
+
+        boolean wasFriends = user.getFriends().remove(friendId);
+
+        return wasFriends;
     }
 
     public List<User> getCommonFriends(int userId, int otherUserId) throws NotFoundException {
         User user = getUserOrThrow(userId);
         User otherUser = getUserOrThrow(otherUserId);
 
-        return user.getFriends().stream()
-                .filter(id -> otherUser.getFriends().contains(id))
+        Set<Integer> commonIds = new HashSet<>(user.getFriends());
+        commonIds.retainAll(otherUser.getFriends());
+
+        return commonIds.stream()
                 .map(id -> {
                     try {
                         return getUserOrThrow(id);
@@ -101,7 +81,8 @@ public class UserService {
     }
 
     public User getUserOrThrow(int id) throws NotFoundException {
-        return Optional.ofNullable(userStorage.findById(id)).orElseThrow(() -> new NotFoundException("Пользователь не найден"));
+        return Optional.ofNullable(userStorage.findById(id))
+                .orElseThrow(() -> new NotFoundException("Пользователь с id=" + id + " не найден"));
     }
 
     public Collection<User> findAll() {
@@ -119,9 +100,12 @@ public class UserService {
         return userStorage.update(user);
     }
 
-
-    public void delete(int id) {
+    public void delete(int id) throws NotFoundException {
+        User user = userStorage.findById(id);
+        if (user == null) {
+            throw new NotFoundException("Пользователь с id=" + id + " не найден");
+        }
         userStorage.delete(id);
+        log.info("Пользователь с id={} удалён", id);
     }
 }
-
